Create wheExpression2.kt
This is a simple Kotlin program that uses the when expression to determine the appropriate action for a given traffic light color.

The program first defines a variable trafficLightColor with the value "Black". Then the 'when' expression is used to evaluate the value of trafficLightColor and execute the corresponding block of code based on the matching case.

In this particular program, since the value of trafficLightColor is "Black", none of the case conditions match. Therefore, the else block is executed and the output will be "Invalid trafficLightColor".

If the value of trafficLightColor was "Red", the program would output "Stop". If it was "Yellow", the output would be "Slow". And if it was "Green", the output would be "Go"