Create IsPrimeNumberIfElse.kt
In this program, we first define the number we want to check for primality, in this case 7. You can replace 7 with any other number between 1 and 10 to check for primality.

The isPrime variable uses a range expression (2 until number) to generate a sequence of integers from 2 up to (but not including) the value of 'number', and then uses the none function to check if none of these integers divide 'number' evenly (i.e., with a remainder of 0). If none of them do, then the isPrime variable is assigned true, indicating that 'number' is prime.

Finally, we output the result of the primality check by using an 'if' expression to check the value of isPrime and the value of number. We include the check for number > 1 to exclude the numbers 1 and 0, which are not considered prime