Create IsPrimeNumberWhen.kt
This is a Kotlin program that uses a when expression to determine if a given number x is a prime number between 1 and 10.

First, a variable x is defined with the value of 3.

Next, the when expression is used to evaluate the value of x and execute the corresponding block of code based on the matching case. In this case, there are four cases defined for the prime numbers 2, 3, 5, and 7.

If the value of x matches one of these cases, the program will output a message indicating that x is a prime number between 1 and 10. Otherwise, the else block is executed, indicating that x is not a prime number between 1 and 10.

In this particular program, since the value of x is 3, the second case is matched, and the output will be "x is a prime number between 1 and 10." If the value of x had been 4 or any other number that is not a prime number between 1 and 10, the else block would have been executed and the output would have been "x isn't a prime number between 1 and 10."