Create safe-call operator.kt
This is a Kotlin program that demonstrates the use of the safe call operator (?.) to access the length property of a nullable string.

First, a variable favouriteActor is defined with the initial value of null. The ? in the variable declaration indicates that favouriteActor is a nullable string variable.

Next, the program uses the safe call operator (?.) to attempt to access the length property of favouriteActor. The safe call operator checks if favouriteActor is null before accessing its length property. If favouriteActor is null, then the program will simply return null instead of throwing a NullPointerException.

In this particular program, since the value of favouriteActor is null, the output will be null. If favouriteActor had contained a non-null string value, the output would have been the length of the string.