Create not-null assertion operator.kt
This is a Kotlin program that demonstrates the use of the not-null assertion operator (!!) to assert that a nullable variable is not null.

First, a variable favouriteActor is defined with the initial value of "Modester". The ? in the variable declaration indicates that favouriteActor is a nullable string variable.

Next, the program uses the not-null assertion operator (!!) to assert that favouriteActor is not null. This operator tells the compiler that we are certain that favouriteActor is not null, and it should not throw a NullPointerException. If favouriteActor is actually null, then a NullPointerException will be thrown at runtime.

After asserting that favouriteActor is not null, the program then accesses the length property of favouriteActor.

In this particular program, since the value of favouriteActor is "Modester", the output will be the length of the string, which is 8. If favouriteActor had contained a null value instead, a NullPointerException would have been thrown at runtime. It's worth noting that the not-null assertion operator should be used with caution, as it can result in runtime errors if used improperly. It's generally recommended to use safe calls (?.) or null checks (if (variable != null) { ... }) instead of the not-null assertion operator