Create Not-null-With-If-statement.kt
This is a Kotlin program that demonstrates how to check if a nullable variable is null and access its properties safely using an if statement.

First, a variable favoriteActor is defined with the initial value of "Modester". The ? in the variable declaration indicates that favoriteActor is a nullable string variable.

Next, the program uses an if statement to check if favoriteActor is not null. If favoriteActor is not null, then the program will execute the code within the if block, which outputs a message indicating the length of the string. The ${} syntax is used to interpolate the length of favoriteActor into the string.

If favoriteActor is null, then the program will execute the code within the else block, which outputs a message indicating that the user didn't input a name.

In this particular program, since the value of favoriteActor is "Modester", the output will be "The number of characters in your favorite actor's name is 8." If favoriteActor had contained a null value instead, the output would have been "You didn't input a name."

Using if statements to check for null values and handle them appropriately is a common practice in Kotlin, especially when working with nullable variables. It's generally recommended to use safe calls (?.) or null checks (if (variable != null) { ... }) to access the properties of nullable variables safely